Remove built-in devices from the main lib
